Fix a MSVC warning.

+++ src/Tools/Algo/Draw_generator/User_pdf_noise_generator/Fast/User_pdf_noise_generator_fast.cpp
@@ -112,7 +112,7 @@ void User_pdf_noise_generator_fast<float>
 	{
 		const auto& cdf_y = signal[i] ? dis.get_cdf_y()[1] : dis.get_cdf_y()[0];
 		const auto& cdf_x = signal[i] ? dis.get_cdf_x()[1] : dis.get_cdf_x()[0];
-		draw[i] = interp_function(cdf_y.data(), cdf_x.data(), cdf_x.size(), draw[i]);
+		draw[i] = interp_function(cdf_y.data(), cdf_x.data(), (unsigned)cdf_x.size(), draw[i]);
 	}
 }
 }
